fix: `ReferenceArrayColumn.serialize` should not perform queries. it should be able to rely on the integrity of `ReferenceArray._ids`.
`ReferenceArray` should maintain synchronization between `_objects` and `_ids`

how the bug was realized:

when using `@proteinjs/user/getScopedDbAsSystem`, the intention is to avoid scope constraints on your db operations. however, when `ReferenceArrayColumn.serialize` had the potential to perform a query, it was not inheriting the lack of scope contraints. instead of trying to find a way to allow `ReferenceArray.get` to inherit this, it was cleaner to just fix the fidelity of `_ids` in `ReferenceArray` and avoid the whole potential query in `ReferenceArrayColumn.serialize`. Besides this issue, it's a performance improvement.